Changed project structure via a Blueprint
