+++ src/server.py
@@ -6,70 +6,15 @@ top-level directory of this distribution.
 from sanic import Sanic, response
 from sanic.response import json
 
-from aiomysql import create_pool
-from aiomysql.sa import create_engine
-
-from sqlalchemy import MetaData, Table, Column, String
-from sqlalchemy.schema import CreateTable
-
 from json import dumps
+
+from listeners import listener_blueprint
 
 
 app = Sanic(__name__)
+app.blueprint(listener_blueprint)
 
-metadata = MetaData()
-table = Table(
-    'links',
-    metadata,
-    Column('endpoint', String(255)),
-    Column('url', String(255))
-)
-
-
-@app.listener('before_server_start')
-async def initialise_db(app, loop):
-    pool = await create_pool(
-        host='db',
-        port=3306,
-        user='user',
-        password='password',
-        db='db',
-        loop=loop,
-        autocommit=True
-    )
-    app.engine = await create_engine(
-        host='db',
-        port=3306,
-        user='user',
-        password='password',
-        db='db',
-        loop=loop
-    )
-    async with pool.acquire() as conn:
-        db_cursor = await conn.cursor()
-        try:
-            await db_cursor.execute(str(CreateTable(table)))
-            query = 'INSERT INTO links (endpoint, url) VALUES (%s, %s)'
-            data = [
-                ('pomuzemesi', 'https://staging.pomuzeme.si'),
-                ('vlk', 'http://www.vlk.cz'),
-                ('kodex', 'https://github.com/Applifting/culture'),
-                ('meta', 'https://github.com/Applifting/link-shortener')
-            ]
-            await db_cursor.executemany(query, data)
-        except Exception as error:
-            print(str(error) + '\n' + 'Table is probably already cached')
-
-        await db_cursor.close()
-
-    pool.terminate()
-    await pool.wait_closed()
-
-
-@app.listener('after_server_stop')
-async def close_engine(app, loop):
-    app.engine.terminate()
-    await app.engine.wait_closed()
+table = listener_blueprint.table
 
 
 @app.route('/api/links', methods=['GET'])
